Refactored pagination, added delete confirmation

+++ app/Http/Controllers/ArticleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Article;
 use App\Http\Requests\ArticleRequest;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Session;
 
@@ -14,33 +15,22 @@ use Illuminate\Support\Facades\Session;
 class ArticleController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a paginated listing of articles.
      *
+     * @param int $currentPage
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        $articles = Article::orderBy('published_on', 'desc')->take(5)->get();
-
-        return view('articles.index', compact('articles'));
-    }
-
-    /**
-     * Display a paginated listing of the resource.
-     *
-     * @param $page
-     * @return \Illuminate\Http\Response
-     */
-    public function paginate($page)
+    public function index($currentPage = 1)
     {
         $perPage = 5;
+        $numberOfArticles = ceil(count(Article::all()) / $perPage);
 
         $articles = Article::orderBy('published_on', 'desc')
-            ->skip($perPage * $page)
+            ->skip($perPage * ($currentPage - 1))
             ->take($perPage)
             ->get();
 
-        return view('articles.index', compact('articles', 'page'));
+        return view('articles.index', compact('articles', 'currentPage', 'numberOfArticles'));
     }
 
     /**
@@ -112,14 +102,26 @@ class ArticleController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param Request $request
      * @param  int $id
      * @return Redirect
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        if (!$request->input('confirmed')) {
+            return redirect(route('articles.confirmDelete', $id));
+        }
+
         Article::findOrFail($id)->delete();
         Session::flash('flash_message', 'Article deleted.');
 
         return redirect(route('articles.index'));
+    }
+
+    public function confirmDelete($id)
+    {
+        $article = Article::findOrFail($id);
+
+        return view('articles.confirm_delete', compact('article'));
     }
 }
